AI/requirements.txt
fastapi
uvicorn
python-docx
pydantic
requests
PyPDF2
